requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alibabacloud-credentials==1.0.2
alibabacloud-credentials-api==1.0.0
alibabacloud-docmind-api20220711==1.4.11
alibabacloud_gateway_spi==0.0.3
alibabacloud-tea==0.4.3
alibabacloud-tea-openapi==0.4.1
alibabacloud_tea_util==0.3.13
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
blake3==1.0.5
cachetools==6.2.0
cbor2==5.7.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
compressed-tensors==0.10.2
crcmod==1.7
cryptography==44.0.3
cupy-cuda12x==13.6.0
darabonba-core==1.0.3
dashscope==1.24.4
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
fastuuid==0.12.0
filelock==3.19.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
gguf==0.17.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
json5==0.12.1
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.2.2
litellm==1.77.1
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistral_common==1.8.5
modelscope==1.30.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
networkx==3.4.2
ninja==1.13.0
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.99.5
openai-harmony==0.0.4
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
outlines_core==0.2.10
packaging==25.0
pandas==2.3.2
partial-json-parser==0.2.1.1.post6
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pip==25.2
prometheus_client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
py-cpuinfo==9.0.0
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pydantic-extra-types==2.10.5
Pygments==2.19.2
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
qwen-agent==0.0.26
ray==2.49.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
safetensors==0.6.2
sandbox-fusion==0.3.7
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.37.1
setproctitle==1.3.7
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.47.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.56.1
triton==3.3.1
typer==0.17.4
typing_extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vllm==0.10.1
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
xformers==0.0.31
xgrammar==0.1.21
xlsxwriter==3.2.8
yarl==1.20.1
zipp==3.23.0
